Remove ListboxItemPanel
Just move the contents into Listbox
