+++ trview.ui/trview.ui.vcxproj
@@ -27,7 +27,6 @@
     <ClInclude Include="Label.h" />
     <ClInclude Include="Listbox.h" />
     <ClInclude Include="ListboxItem.h" />
-    <ClInclude Include="ListboxItemPanel.h" />
     <ClInclude Include="NumericUpDown.h" />
     <ClInclude Include="SizeDimension.h" />
     <ClInclude Include="SizeMode.h" />
@@ -44,7 +43,6 @@
     <ClCompile Include="Label.cpp" />
     <ClCompile Include="Listbox.cpp" />
     <ClCompile Include="ListboxItem.cpp" />
-    <ClCompile Include="ListboxItemPanel.cpp" />
     <ClCompile Include="NumericUpDown.cpp" />
     <ClCompile Include="Slider.cpp" />
     <ClCompile Include="StackPanel.cpp" />
